auto

+++ rtp_llm/models_py/modules/moe/executors/test/trtllm_fp4_executor_test.py
@@ -226,7 +226,7 @@ def _generate_ref_output(
         device_support_pdl(payload.expert_x.device),
         GatedActType.SwiGlu.value,  # gated_act_type
         None,
-    )[0].to(torch.float)
+    )[0]
 
     print(output)
     print(ref_output)
